requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
loguru==0.7.2
diffusers
transformers
accelerate
xformers==0.0.27.post2
httpx==0.27.2
pillow==10.4.0
numpy==2.0.2
omegaconf==2.3.0
trimesh==4.0.5
huggingface-hub
imageio[ffmpeg]
onnxruntime
python-multipart
hf-transfer
sentencepiece
python-dotenv
ninja
pybind11
einops
opencv-python
tqdm
pymeshlab
pygltflib
xatlas